+++ spec/swagger/api/v1/photos_controller_spec.rb
@@ -111,6 +111,7 @@ RSpec.describe 'Api::V1::PhotosController', type: :request do
                    state: { type: :string },
                    country: { type: :string },
                    type: { type: :string },
+                   orientation: { type: :string },
                    source: { type: :string }
                  },
                  required: %w[id latitude longitude localDateTime originalFileName city state country type source]
@@ -143,7 +144,8 @@ RSpec.describe 'Api::V1::PhotosController', type: :request do
                  state: { type: :string },
                  country: { type: :string },
                  type: { type: :string },
-                 source: { type: :string }
+                 source: { type: :string },
+                 orientation: { type: :string, enum: %w[portrait landscape] }
                }
 
         let(:id) { '7fe486e3-c3ba-4b54-bbf9-1281b39ed15c' }